Move validate_custom_parameters into validators directory
