+++ tests/agent_features/test_browser.py
@@ -18,9 +18,9 @@ import json
 
 import six
 
-from testing_support.fixtures import (override_application_settings,
-    validate_custom_parameters)
+from testing_support.fixtures import override_application_settings
 from testing_support.validators.validate_transaction_errors import validate_transaction_errors
+from testing_support.validators.validate_custom_parameters import validate_custom_parameters
 from newrelic.api.application import application_settings
 from newrelic.api.transaction import (get_browser_timing_header,
         get_browser_timing_footer, add_custom_parameter,
