+++ tests/agent_features/test_high_security_mode.py
@@ -24,11 +24,11 @@ from testing_support.fixtures import (
     validate_attributes_complete,
     validate_custom_event_count,
     validate_custom_event_in_application_stats_engine,
-    validate_custom_parameters,
     validate_non_transaction_error_event,
     validate_request_params_omitted,
     validate_tt_segment_params,
 )
+from testing_support.validators.validate_custom_parameters import validate_custom_parameters
 from testing_support.validators.validate_transaction_errors import validate_transaction_errors
 from newrelic.api.application import application_instance as application
 from newrelic.api.background_task import background_task
